requirements-test.txt
coverage==4.1.0
django-markdown-deux==1.0.5
psycopg2==2.6.2
